removed unused fonts, added ingredeintMap
