--- src/constants/IngredientsToExclude.tsx
@@ -1,24 +1,0 @@
-export const INGS_TO_EXCLUDE = [
-  "salt",
-  "pepper",
-  "sugar",
-  "water",
-  "flour",
-  "baking",
-  "cream",
-  "vanilla",
-  "ice cubes",
-  "mayonnaise",
-  "mustard",
-  "ketchup",
-  "butter",
-  "density",
-  "cooking oil",
-  "vegetable oil",
-  "olive oil",
-  "extract",
-  "nutmeg",
-  "gelatin",
-  "egg substitute",
-  "half n half"
-];
