+++ src/screens/SavoredListScreen.tsx
@@ -34,6 +34,7 @@ export default function SavoredListScreen({
 }: SavoredListScreenProps) {
   const [fontsLoaded] = useFonts({
     OpenSans: require("../../assets/fonts/OpenSans-Regular.ttf"),
+    OpenSansBold: require("../../assets/fonts/OpenSans-Bold.ttf"),
     Satisfy: require("../../assets/fonts/Satisfy-Regular.ttf"),
   });
